Update .nuspec
removed release notes as nuget's website does not really support proper release
notes (and we are probably too lazy to maintain outside release notes.)

Also simplified authors to just miaplaza inc. The curious reader can have a
look at the git history but it's just hard to say when somebody should be
considered an author and when not.